+++ game/systems/s_tilemap.cpp
@@ -7,9 +7,11 @@
 #include <array>
 
 #include "log.h"
+#include "input.h"
 
 #include "r_pipeline.h"
 #include "r_renderer.h"
+#include "r_window.h"
 #include "stb_image.h"
 
 using namespace Verse;
@@ -39,13 +41,11 @@ std::vector<std::vector<ui8>> System::Tilemap::load(str path) {
 
 void System::Tilemap::init(Config &c) {
     for (EntityID e : SceneView<Component::Tilemap>(*c.active_scene)) {
-        createVertices(c, e);
+        createVertices(c, c.active_scene->getComponent<Component::Tilemap>(e));
     }
 }
 
-void System::Tilemap::createVertices(Config &c, EntityID eid) {
-    Component::Tilemap* tmap = c.active_scene->getComponent<Component::Tilemap>(eid);
-    
+void System::Tilemap::createVertices(Config &c, Component::Tilemap* tmap) {
     tmap->vert = {};
     for (int l = 0; l < tmap->tex_id.size(); l++) {
         tmap->vert.push_back({});
@@ -91,6 +91,9 @@ void System::Tilemap::render(Config &c) {
     for (EntityID e : SceneView<Component::Tilemap>(*c.active_scene)) {
         Component::Tilemap* tmap = c.active_scene->getComponent<Component::Tilemap>(e);
         
+        if (c.tilemap_editor and tmap != c.current_tilemap_edit)
+            continue;
+        
         int i = 0;
         for (ui32 t : tmap->tex_id) {
             Graphics::Renderer::renderTilemap(c, t, reinterpret_cast<float*>(tmap->vert[i].data()), (int)tmap->vert[i].size(), tmap->layer);
@@ -106,4 +109,32 @@ Vec2 System::Tilemap::calculateSize(Component::Tilemap* tmap) {
     size.y = (int)(tmap->tiles.size() * tmap->tex_size.y);
     
     return size;
+}
+
+void System::Tilemap::renderEditor(Config &c) {
+    Vec2 t_size = calculateSize(c.current_tilemap_edit) - c.current_tilemap_edit->tex_size;
+    Vec2 t_pos = c.current_tilemap_edit->pos;
+    
+    Vec2 m_pos = Graphics::Window::windowToScene(c, Input::mouse());
+    
+    Vec2 rel_pos = m_pos - t_pos;
+    
+    if (rel_pos.x >= 0 and rel_pos.y >= 0 and rel_pos.x < t_size.x and rel_pos.y < t_size.y) {
+        Vec2 curr_tile = Vec2(floor((float)rel_pos.x / (float)c.current_tilemap_edit->tex_size.x),
+                              floor((float)rel_pos.y / (float)c.current_tilemap_edit->tex_size.y));
+        Vec2 curr_pos = t_pos + Vec2(curr_tile.x * c.current_tilemap_edit->tex_size.x,
+                                     curr_tile.y * c.current_tilemap_edit->tex_size.y);
+        
+        Graphics::Renderer::renderDebugCollider(c, Rect2(curr_pos, c.current_tilemap_edit->tex_size), false);
+        
+        if (Input::down((ui8)Input::MouseButton::Left)) {
+            c.current_tilemap_edit->tiles[curr_tile.y][curr_tile.x] = 0;
+            createVertices(c, c.current_tilemap_edit);
+        }
+        
+        if (Input::down((ui8)Input::MouseButton::Right)) {
+            c.current_tilemap_edit->tiles[curr_tile.y][curr_tile.x] = 255;
+            createVertices(c, c.current_tilemap_edit);
+        }
+    }
 }