+++ game/systems/s_tilemap.h
@@ -10,7 +10,9 @@ namespace Verse::System::Tilemap
 {
     std::vector<std::vector<ui8>> load(str path);
     void init(Config &c);
-    void createVertices(Config &c, EntityID eid);
+    void createVertices(Config &c, Component::Tilemap* tmap);
     void render(Config &c);
     Vec2 calculateSize(Component::Tilemap* tmap);
+
+    void renderEditor(Config &c);
 }